javascriptのリファクタリングとflashの表示
