+++ app/controllers/articles_controller.rb
@@ -15,7 +15,7 @@ class ArticlesController < ApplicationController
 
   def create
     @article = current_user.articles.build(article_params)
-    if @article.save!
+    if @article.save
       redirect_to articles_path, notice: '保存されました！'
     else
       flash.now[:error] = '保存に失敗しました'
