+++ app/models/article.rb
@@ -18,6 +18,9 @@
 #  fk_rails_...  (user_id => users.id)
 #
 class Article < ApplicationRecord
+  validates :title, presence: true
+  validates :content, presence: true
+
   belongs_to :user
   
   has_many :comments, dependent: :destroy
