+++ app/models/comment.rb
@@ -15,6 +15,8 @@
 #  index_comments_on_user_id     (user_id)
 #
 class Comment < ApplicationRecord
+  validates :content, presence: true
+
   belongs_to :article
   belongs_to :user
