+++ app/views/articles/show.html.haml
@@ -44,8 +44,5 @@
       = button_tag type: :submit do
         %i.far.fa-paper-plane.fa-2x
   .comments-container
-    -# .comment-card
-    -# = image_tag
-    -#   .comment-card-content
-    -#     .comment-card-info-username
-    -#     .comment-card-info-content
+-# JavaScriptの読み込み
+= javascript_pack_tag 'article'
