+++ app/views/layouts/application.html.haml
@@ -11,6 +11,8 @@
     %link{:href => "https://use.fontawesome.com/releases/v5.9.0/css/all.css", :rel => "stylesheet"}/
 
   %body
-    -# %p.notice= notice
-    -# %p.alert= alert
     = yield
+    - if flash.present?
+      .flash
+        - flash.each do |key,value|
+          %div{:class => key}= value
